第 35 轮回归成果：权威 md 802→776（累计 3324→776；relay/community 归零，guild -16、dbmw/monitor 各 -3；解析器修复 stun/auction/point 假回归）

+++ dnf_decompile/source/DNFServer/GameServer/Guild/GuildCargo.cpp
@@ -356,12 +356,12 @@ void CGuildCargo::SetGuildCargoDBInfo(STGuildCargoDBInfo& info)
 
 void CGuildCargo::SetGuildCargoHistory(unsigned int idx, STGuildCargoLog* log)
 {
-    int c = (int)idx;
+    unsigned int c = idx;
     if (c > 0x32)
     {
         c = 0x32;
     }
-    for (int i = 0; i < c; i++)
+    for (int i = 0; i < (int)c; i++)
     {
         m_history.push_front(log[i]);
     }
